Add standalone web app with npx support (#23)
