+++ .gitignore
@@ -9,6 +9,7 @@ dist
 extension/README.md
 scripts/.dev-relay.js
 scripts/.dev-relay.js.map
+app/dist/
 next-env.d.ts
 *.tsbuildinfo
