+++ app/src/server.ts
@@ -1,0 +1,66 @@
+/**
+ * Combined HTTP server: serves the visualizer UI and streams events via SSE.
+ * Reuses the extension's hook server, transcript parser, and session watcher.
+ */
+import * as http from 'http'
+import { exec, execFile } from 'child_process'
+
+import { createRelay } from '../../scripts/relay'
+import { serveStatic } from './static'
+
+interface ServerOptions {
+  port: number
+  openBrowser: boolean
+  workspace: string
+  verbose?: boolean
+}
+
+export async function startServer(options: ServerOptions) {
+  const { port, openBrowser, workspace } = options
+
+  const relay = await createRelay({ workspace, verbose: options.verbose })
+
+  const server = http.createServer((req, res) => {
+    // SSE endpoint
+    if (req.url === '/events') {
+      return relay.handleSSE(req, res)
+    }
+
+    // Static files (UI)
+    if (req.method === 'GET') {
+      return serveStatic(req, res)
+    }
+
+    res.writeHead(404)
+    res.end('Not found')
+  })
+
+  server.listen(port, '127.0.0.1', () => {
+    const url = `http://127.0.0.1:${port}`
+    console.log(`Server running at ${url}`)
+    console.log('Waiting for agent events...\n')
+
+    if (openBrowser) {
+      openURL(url)
+    }
+  })
+
+  // Cleanup on exit
+  function cleanup() {
+    server.close()
+    relay.dispose()
+    process.exit(0)
+  }
+  process.on('SIGINT', cleanup)
+  process.on('SIGTERM', cleanup)
+}
+
+function openURL(url: string) {
+  if (process.platform === 'win32') {
+    // 'start' is a shell builtin on Windows — must use exec, not execFile
+    exec(`start "" "${url}"`)
+  } else {
+    const cmd = process.platform === 'darwin' ? 'open' : 'xdg-open'
+    execFile(cmd, [url], () => {})
+  }
+}
