+++ package.json
@@ -10,7 +10,8 @@
     "build:extension": "pnpm --filter agent-flow run build",
     "build:web": "pnpm --filter agent-flow-web run build",
     "build:webview": "pnpm --filter agent-flow-web run build:webview",
-    "build:all": "pnpm run build:webview && pnpm run build:extension"
+    "build:all": "pnpm run build:webview && pnpm run build:extension",
+    "build:app": "node app/build.js"
   },
   "devDependencies": {
     "concurrently": "^9.2.1",
