+++ scripts/relay.ts
@@ -1,0 +1,424 @@
+/**
+ * Shared relay module — receives agent events and streams them to SSE clients.
+ * Used by both the dev relay server and the standalone app.
+ */
+import * as http from 'http'
+import * as crypto from 'crypto'
+import * as fs from 'fs'
+import * as path from 'path'
+import * as os from 'os'
+
+import { HookServer } from '../extension/src/hook-server'
+import { AgentEvent, SessionInfo, WatchedSession } from '../extension/src/protocol'
+import { TranscriptParser } from '../extension/src/transcript-parser'
+import { readNewFileLines } from '../extension/src/fs-utils'
+import { scanSubagentsDir, readSubagentNewLines } from '../extension/src/subagent-watcher'
+import { handlePermissionDetection } from '../extension/src/permission-detection'
+import {
+  INACTIVITY_TIMEOUT_MS, SCAN_INTERVAL_MS, ACTIVE_SESSION_AGE_S, POLL_FALLBACK_MS,
+  SESSION_ID_DISPLAY, SYSTEM_PROMPT_BASE_TOKENS, ORCHESTRATOR_NAME,
+  HOOK_SERVER_NOT_STARTED, WORKSPACE_HASH_LENGTH,
+} from '../extension/src/constants'
+import { setLogLevel } from '../extension/src/logger'
+
+const MAX_EVENT_BUFFER = 5000
+const DISCOVERY_DIR = path.join(os.homedir(), '.claude', 'agent-flow')
+const CLAUDE_DIR = path.join(os.homedir(), '.claude', 'projects')
+
+let relayCreated = false
+let verbose = false
+
+function log(...args: unknown[]) {
+  if (verbose) console.log(...args)
+}
+
+// ─── SSE client management ──────────────────────────────────────────────────
+
+const sseClients = new Set<http.ServerResponse>()
+
+function sendSSE(res: http.ServerResponse, data: unknown) {
+  try { res.write(`data: ${JSON.stringify(data)}\n\n`) } catch {
+    sseClients.delete(res)
+  }
+}
+
+function broadcast(data: string) {
+  for (const res of sseClients) {
+    try { res.write(`data: ${data}\n\n`) } catch {
+      sseClients.delete(res)
+    }
+  }
+}
+
+// ─── Event buffering ────────────────────────────────────────────────────────
+
+const eventBuffer = new Map<string, AgentEvent[]>()
+
+function broadcastEvent(event: AgentEvent) {
+  const sid = event.sessionId?.slice(0, SESSION_ID_DISPLAY) || '?'
+  log(`[event] ${event.type} (session ${sid})`)
+
+  if (event.sessionId) {
+    let buf = eventBuffer.get(event.sessionId) || []
+    buf.push(event)
+    if (buf.length > MAX_EVENT_BUFFER) {
+      buf = buf.slice(buf.length - MAX_EVENT_BUFFER)
+    }
+    eventBuffer.set(event.sessionId, buf)
+  }
+
+  broadcast(JSON.stringify({ type: 'agent-event', event }))
+}
+
+function broadcastSessionLifecycle(type: 'started' | 'ended' | 'updated', sessionId: string, label: string) {
+  if (type === 'started') {
+    broadcast(JSON.stringify({
+      type: 'session-started',
+      session: { id: sessionId, label, status: 'active', startTime: Date.now(), lastActivityTime: Date.now() } as SessionInfo,
+    }))
+  } else if (type === 'ended') {
+    broadcast(JSON.stringify({ type: 'session-ended', sessionId }))
+  } else if (type === 'updated') {
+    broadcast(JSON.stringify({ type: 'session-updated', sessionId, label }))
+  }
+}
+
+// ─── Session watcher ────────────────────────────────────────────────────────
+
+const sessions = new Map<string, WatchedSession>()
+
+function elapsed(sessionId?: string): number {
+  if (sessionId) {
+    const session = sessions.get(sessionId)
+    if (session) return (Date.now() - session.sessionStartTime) / 1000
+  }
+  return 0
+}
+
+function emitContextUpdate(agentName: string, session: WatchedSession, sessionId?: string) {
+  const bd = session.contextBreakdown
+  const total = bd.systemPrompt + bd.userMessages + bd.toolResults + bd.reasoning + bd.subagentResults
+  broadcastEvent({
+    time: elapsed(sessionId),
+    type: 'context_update',
+    payload: { agent: agentName, tokens: total, breakdown: { ...bd } },
+    sessionId,
+  })
+}
+
+function emitEvent(event: AgentEvent, sessionId?: string) {
+  broadcastEvent(sessionId ? { ...event, sessionId } : event)
+}
+
+const parser = new TranscriptParser({
+  emit: emitEvent,
+  elapsed,
+  getSession: (sessionId: string) => sessions.get(sessionId),
+  fireSessionLifecycle: (event) => broadcastSessionLifecycle(event.type, event.sessionId, event.label),
+  emitContextUpdate,
+})
+
+const watcherDelegate = {
+  emit: emitEvent,
+  elapsed,
+  getSession: (sessionId: string) => sessions.get(sessionId),
+  getLastActivityTime: (sessionId: string) => sessions.get(sessionId)?.lastActivityTime,
+  resetInactivityTimer: (sessionId: string) => resetInactivityTimer(sessionId),
+}
+
+function resetInactivityTimer(sessionId: string) {
+  const session = sessions.get(sessionId)
+  if (!session) return
+
+  const wasCompleted = session.sessionCompleted
+  session.lastActivityTime = Date.now()
+  session.sessionCompleted = false
+
+  if (wasCompleted) {
+    broadcastEvent({
+      time: elapsed(sessionId),
+      type: 'agent_spawn',
+      payload: { name: ORCHESTRATOR_NAME, isMain: true, task: session.label, ...(session.model ? { model: session.model } : {}) },
+      sessionId,
+    })
+    broadcastSessionLifecycle('started', sessionId, session.label)
+  }
+
+  if (session.inactivityTimer) clearTimeout(session.inactivityTimer)
+  session.inactivityTimer = setTimeout(() => {
+    if (!session.sessionCompleted && session.sessionDetected) {
+      log(`[session] ${sessionId.slice(0, SESSION_ID_DISPLAY)} inactive`)
+      session.sessionCompleted = true
+      broadcastEvent({
+        time: elapsed(sessionId),
+        type: 'agent_complete',
+        payload: { name: ORCHESTRATOR_NAME },
+        sessionId,
+      })
+      broadcastSessionLifecycle('ended', sessionId, session.label)
+    }
+  }, INACTIVITY_TIMEOUT_MS)
+}
+
+function watchSession(sessionId: string, filePath: string) {
+  const defaultLabel = `Session ${sessionId.slice(0, SESSION_ID_DISPLAY)}`
+  const session: WatchedSession = {
+    sessionId, filePath,
+    fileWatcher: null, pollTimer: null, fileSize: 0,
+    sessionStartTime: Date.now(),
+    pendingToolCalls: new Map(),
+    seenToolUseIds: new Set(),
+    seenMessageHashes: new Set(),
+    sessionDetected: false, sessionCompleted: false,
+    lastActivityTime: Date.now(),
+    inactivityTimer: null,
+    subagentWatchers: new Map(),
+    spawnedSubagents: new Set(),
+    inlineProgressAgents: new Set(),
+    subagentsDirWatcher: null, subagentsDir: null,
+    label: defaultLabel, labelSet: false,
+    model: null,
+    permissionTimer: null, permissionEmitted: false,
+    contextBreakdown: { systemPrompt: SYSTEM_PROMPT_BASE_TOKENS, userMessages: 0, toolResults: 0, reasoning: 0, subagentResults: 0 },
+  }
+  sessions.set(sessionId, session)
+
+  const stat = fs.statSync(filePath)
+  const catchUpEntries = parser.prescanExistingContent(filePath, stat.size, session)
+  session.fileSize = stat.size
+  parser.extractSessionLabel(catchUpEntries, session)
+
+  broadcastSessionLifecycle('started', sessionId, session.label)
+  broadcastEvent({
+    time: 0, type: 'agent_spawn',
+    payload: { name: ORCHESTRATOR_NAME, isMain: true, task: session.label, ...(session.model ? { model: session.model } : {}) },
+    sessionId,
+  })
+  session.sessionDetected = true
+
+  emitContextUpdate(ORCHESTRATOR_NAME, session, sessionId)
+  parser.emitCatchUpEntries(catchUpEntries, session, sessionId)
+
+  session.fileWatcher = fs.watch(filePath, (eventType) => {
+    if (eventType === 'change') readNewLines(sessionId)
+  })
+
+  session.pollTimer = setInterval(() => {
+    readNewLines(sessionId)
+    for (const [subPath] of session.subagentWatchers) {
+      readSubagentNewLines(watcherDelegate, parser, subPath, sessionId)
+    }
+    scanSubagentsDir(watcherDelegate, parser, sessionId)
+  }, POLL_FALLBACK_MS)
+
+  session.subagentsDir = path.join(path.dirname(filePath), sessionId, 'subagents')
+  scanSubagentsDir(watcherDelegate, parser, sessionId)
+  resetInactivityTimer(sessionId)
+
+  log(`[session] Watching ${sessionId.slice(0, SESSION_ID_DISPLAY)} — "${session.label}"`)
+}
+
+function readNewLines(sessionId: string) {
+  const session = sessions.get(sessionId)
+  if (!session) return
+
+  const result = readNewFileLines(session.filePath, session.fileSize)
+  if (!result) return
+  session.fileSize = result.newSize
+  for (const line of result.lines) {
+    parser.processTranscriptLine(line, ORCHESTRATOR_NAME, session.pendingToolCalls, session.seenToolUseIds, sessionId, session.seenMessageHashes)
+  }
+
+  handlePermissionDetection(watcherDelegate, ORCHESTRATOR_NAME, session.pendingToolCalls, session, sessionId, session.sessionCompleted, true)
+  scanSubagentsDir(watcherDelegate, parser, sessionId)
+  resetInactivityTimer(sessionId)
+}
+
+// ─── Session scanner ────────────────────────────────────────────────────────
+
+function scanForActiveSessions(workspace: string) {
+  if (!fs.existsSync(CLAUDE_DIR)) return
+
+  let resolved = workspace
+  try { resolved = fs.realpathSync(resolved) } catch {}
+  const encoded = resolved.replace(/[/\\:]/g, '-')
+
+  const dirsToScan: string[] = []
+  const projectDir = path.join(CLAUDE_DIR, encoded)
+  if (fs.existsSync(projectDir)) dirsToScan.push(projectDir)
+
+  try {
+    for (const dir of fs.readdirSync(CLAUDE_DIR, { withFileTypes: true })) {
+      if (!dir.isDirectory()) continue
+      const fullPath = path.join(CLAUDE_DIR, dir.name)
+      if (fullPath === projectDir) continue
+      if (dir.name.startsWith(encoded + '-')) {
+        dirsToScan.push(fullPath)
+      }
+    }
+  } catch {}
+
+  for (const dirPath of dirsToScan) {
+    try {
+      for (const file of fs.readdirSync(dirPath)) {
+        if (!file.endsWith('.jsonl')) continue
+        const filePath = path.join(dirPath, file)
+        const stat = fs.statSync(filePath)
+        const sessionId = path.basename(file, '.jsonl')
+
+        let newestMtime = stat.mtimeMs
+        const subagentsDir = path.join(dirPath, sessionId, 'subagents')
+        try {
+          if (fs.existsSync(subagentsDir)) {
+            for (const subFile of fs.readdirSync(subagentsDir)) {
+              if (!subFile.endsWith('.jsonl')) continue
+              const subStat = fs.statSync(path.join(subagentsDir, subFile))
+              if (subStat.mtimeMs > newestMtime) newestMtime = subStat.mtimeMs
+            }
+          }
+        } catch {}
+
+        const ageSeconds = (Date.now() - newestMtime) / 1000
+        if (ageSeconds <= ACTIVE_SESSION_AGE_S && !sessions.has(sessionId)) {
+          watchSession(sessionId, filePath)
+        }
+      }
+    } catch {}
+  }
+}
+
+// ─── Discovery file ─────────────────────────────────────────────────────────
+
+function normalizePath(p: string): string {
+  let resolved = path.resolve(p)
+  try { resolved = fs.realpathSync(resolved) } catch {}
+  return resolved
+}
+
+function hashWorkspace(workspace: string): string {
+  return crypto.createHash('sha256').update(normalizePath(workspace)).digest('hex').slice(0, WORKSPACE_HASH_LENGTH)
+}
+
+let discoveryFilePath: string | null = null
+
+function writeDiscoveryFile(port: number, workspace: string) {
+  if (!fs.existsSync(DISCOVERY_DIR)) fs.mkdirSync(DISCOVERY_DIR, { recursive: true })
+  const hash = hashWorkspace(workspace)
+  discoveryFilePath = path.join(DISCOVERY_DIR, `${hash}-${process.pid}.json`)
+  fs.writeFileSync(discoveryFilePath, JSON.stringify({ port, pid: process.pid, workspace: normalizePath(workspace) }, null, 2) + '\n')
+}
+
+function removeDiscoveryFile() {
+  if (discoveryFilePath) {
+    try { fs.unlinkSync(discoveryFilePath) } catch {}
+  }
+}
+
+// ─── Public API ─────────────────────────────────────────────────────────────
+
+export interface Relay {
+  /** Handle an incoming SSE connection */
+  handleSSE: (req: http.IncomingMessage, res: http.ServerResponse) => void
+  /** Clean up all resources */
+  dispose: () => void
+}
+
+export interface RelayOptions {
+  workspace: string
+  verbose?: boolean
+}
+
+export async function createRelay(options: RelayOptions): Promise<Relay> {
+  const { workspace } = options
+  verbose = options.verbose ?? false
+  if (!verbose) setLogLevel('error')
+  if (relayCreated) {
+    throw new Error('createRelay() can only be called once per process')
+  }
+  relayCreated = true
+
+  const hookServer = new HookServer()
+  const hookPort = await hookServer.start()
+  if (hookPort === HOOK_SERVER_NOT_STARTED) {
+    throw new Error('Failed to start hook server (port in use)')
+  }
+
+  hookServer.onEvent((event: AgentEvent) => {
+    broadcast(JSON.stringify({ type: 'agent-event', event }))
+  })
+
+  writeDiscoveryFile(hookPort, workspace)
+
+  scanForActiveSessions(workspace)
+  const scanInterval = setInterval(() => scanForActiveSessions(workspace), SCAN_INTERVAL_MS)
+
+  const resolved = (() => { try { return fs.realpathSync(workspace) } catch { return workspace } })()
+  const encoded = resolved.replace(/[/\\:]/g, '-')
+  const projectDir = path.join(CLAUDE_DIR, encoded)
+  let projectDirWatcher: fs.FSWatcher | null = null
+  if (fs.existsSync(projectDir)) {
+    try {
+      projectDirWatcher = fs.watch(projectDir, (_eventType, filename) => {
+        if (filename?.endsWith('.jsonl')) scanForActiveSessions(workspace)
+      })
+    } catch {}
+  }
+
+  return {
+    handleSSE(req: http.IncomingMessage, res: http.ServerResponse) {
+      res.writeHead(200, {
+        'Content-Type': 'text/event-stream',
+        'Cache-Control': 'no-cache',
+        'Connection': 'keep-alive',
+      })
+
+      sseClients.add(res)
+      log(`[sse] Client connected (${sseClients.size} total)`)
+
+      req.on('close', () => {
+        sseClients.delete(res)
+        log(`[sse] Client disconnected (${sseClients.size} total)`)
+      })
+
+      // Send current session list
+      const sessionList: SessionInfo[] = []
+      for (const session of sessions.values()) {
+        if (!session.sessionDetected) continue
+        sessionList.push({
+          id: session.sessionId, label: session.label,
+          status: session.sessionCompleted ? 'completed' : 'active',
+          startTime: session.sessionStartTime, lastActivityTime: session.lastActivityTime,
+        })
+      }
+      if (sessionList.length > 0) {
+        sendSSE(res, { type: 'session-list', sessions: sessionList })
+      }
+
+      // Replay buffered events for the most recent active session
+      const sorted = [...sessionList].sort((a, b) => {
+        const aActive = a.status === 'active' ? 1 : 0
+        const bActive = b.status === 'active' ? 1 : 0
+        if (aActive !== bActive) return bActive - aActive
+        return b.lastActivityTime - a.lastActivityTime
+      })
+      if (sorted.length > 0) {
+        const buffered = eventBuffer.get(sorted[0].id)
+        if (buffered) {
+          sendSSE(res, { type: 'agent-event-batch', events: buffered })
+        }
+      }
+    },
+
+    dispose() {
+      removeDiscoveryFile()
+      hookServer.dispose()
+      clearInterval(scanInterval)
+      projectDirWatcher?.close()
+      for (const session of sessions.values()) {
+        session.fileWatcher?.close()
+        if (session.pollTimer) clearInterval(session.pollTimer)
+        if (session.inactivityTimer) clearTimeout(session.inactivityTimer)
+      }
+    },
+  }
+}
