+++ web/app-entry.tsx
@@ -1,0 +1,9 @@
+import { createRoot } from 'react-dom/client'
+import { AgentVisualizer } from './components/agent-visualizer'
+import './app/globals.css'
+
+// Standalone CLI entry — no VS Code API, connects to relay via SSE
+const rootElement = document.getElementById('root')
+if (!rootElement) throw new Error('Root element #root not found')
+const root = createRoot(rootElement)
+root.render(<AgentVisualizer />)
