+++ web/hooks/use-vscode-bridge.ts
@@ -69,11 +69,12 @@ export function useVSCodeBridge(): BridgeHookResult {
     // Skip in VS Code — extension handles events via postMessage
     if (bridge.isVSCode) return
 
-    // Only connect in development mode without demo
-    if (process.env.NODE_ENV !== 'development' || process.env.NEXT_PUBLIC_DEMO !== '0') return
+    // Connect to relay in dev mode or standalone CLI mode
+    const isStandalone = process.env.AGENT_FLOW_STANDALONE === '1'
+    if (!isStandalone && (process.env.NODE_ENV !== 'development' || process.env.NEXT_PUBLIC_DEMO !== '0')) return
 
-    const relayPort = process.env.NEXT_PUBLIC_RELAY_PORT || '3001'
-    const es = new EventSource(`http://127.0.0.1:${relayPort}/events`)
+    const relayPort = process.env.NEXT_PUBLIC_RELAY_PORT || ''
+    const es = new EventSource(relayPort ? `http://127.0.0.1:${relayPort}/events` : '/events')
 
     es.onopen = () => {
       setConnectionStatus('connected')
